Improvements AND Clean code

+++ DownloadUtils.cs
@@ -1,21 +1,21 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace YoutubeDownloader {
+﻿namespace YoutubeDownloader {
     internal class DownloadUtils {
-        public static void Download(DownloadPreferences downloadPreferences, Utils.EventOutputCommand eventOutput) {
+        public static void Download(DownloadPreferences? downloadPreferences, Utils.EventOutputCommand eventOutput) {
+            if (downloadPreferences == null) { return; }
+
+
             Utils.RunCommand($"yt-dlp.exe -f {downloadPreferences.FormatString} -o {downloadPreferences.media.ID}.%(ext)s {downloadPreferences.media.URL}", eventOutput);
 
-            if (File.Exists(downloadPreferences.path)) {
-                File.Delete(downloadPreferences.path);
+            if (downloadPreferences.path != null) {
+                if (File.Exists(downloadPreferences.path)) {
+                    File.Delete(downloadPreferences.path);
+                }
+
+                if (File.Exists($"{downloadPreferences.media.ID}.{downloadPreferences.FileExt}")) {
+                    File.Move($"{downloadPreferences.media.ID}.{downloadPreferences.FileExt}", downloadPreferences.path);
+                }
             }
 
-            if (File.Exists($"{downloadPreferences.media.ID}.{downloadPreferences.FileExt}")) {
-                File.Move($"{downloadPreferences.media.ID}.{downloadPreferences.FileExt}", downloadPreferences.path);
-            }
         }
 
     }
